Search using the common library
This commit removes the local searching, since the common library will be doing that for us from now on.
Also I removed the custom JSON [de]serializing code, since we do not need to cache the lists anymore. Because of these changes, we were able to clean up quite some code on the Android side