Add Cleanup call, fix some issues with OpenVPN status updates
Adds the cleanup call, and fixes some incorrect 'disconnect' status notifications in the UI and backend.
This was mostly due to the UI requesting a status update immediately from the ViewModel, which then triggered a status update with the disconnected state.
Another issue was an incorrect VPN level to status mapping, where we incorrectly treated the 'no network' level as a disconnected state, while actually it was connecting.