BackendService: Fix ProxyGuard source port not set to 0
The sourceport should only be set in the Linux client as that can
re-use source ports. If we set the source port from common (this was
not documented well by me), the Android client will not be able to
reconnect.

Let's set the source port to zero such that the OS can choose an
available source port for reconnecting